refactor(#1120): fold review fixes — types.ts cap headroom, baseline, ::date
- Revert DEPRECATED_WEB_TREE_BASELINE 226->170: --update-baseline wrongly
  bumped the non-failing tree ratchet, absorbing ~55 files of unrelated drift.
  Keep only the intended reach-in 161->162.
- Extract VehicleClassRepository/Filters into types-vehicle-class.ts (mirrors
  types-vehicle-block split); types.ts 798->788 so the develop rebase clears
  the 800 hard cap. Drop the now-orphaned internal VehicleClass import.
- Cast compliance date params ::date (drizzle/vehicle.ts), matching the
  booking cutoff's explicit cast — driver-robustness + sibling consistency.
