+++ scripts/lint-module-boundaries.ts
@@ -100,7 +100,7 @@ function isForbiddenWebDbSpec(spec: string): boolean {
 //
 // After a drain, refresh in place with:
 //   bun run scripts/lint-module-boundaries.ts --update-baseline
-const DEPRECATED_WEB_TREE_BASELINE = 226
+const DEPRECATED_WEB_TREE_BASELINE = 170
 
 // #1110: cross-feature reach-ins into vite/<feature>/<internal> from outside
 // that feature. Counted globally; the rule is a ratchet (monotonic
